Add component logic to seperate components
