+++ projects/player/src/app/components/interaction-phonetics/interaction-phonetics.component.scss
@@ -2,10 +2,8 @@
 @use '_variables.scss';
 
 .phonetics-wrapper {
+  @include variables.flex-center;
+  flex-direction: row;
   width: 100%;
   height: 150px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
 }
