+++ projects/player/src/app/components/interaction-syllabify/interaction-syllabify.component.scss
@@ -1,0 +1,9 @@
+@use '_colors.scss';
+@use '_variables.scss';
+
+.syllabify-wrapper {
+  @include variables.flex-center;
+  flex-direction: row;
+  width: 100%;
+  height: 150px;
+}
