+++ projects/player/src/app/components/interaction-syllabify/interaction-syllabify.component.ts
@@ -1,9 +1,11 @@
 import {
   Component, input, OnInit, signal
 } from '@angular/core';
+import { Response } from '@iqbspecs/response/response.interface';
 import { InteractionComponentDirective } from '../../directives/interaction-component.directive';
 import { NumberedOption, SyllabifyParams } from '../../models/unit-definition';
 import { StandardButtonComponent } from '../../shared/standard-button/standard-button.component';
+import { createNumberedOptions } from '../../utils/option-helpers';
 
 @Component({
   selector: 'stars-interaction-syllabify',
@@ -18,27 +20,35 @@ import { StandardButtonComponent } from '../../shared/standard-button/standard-b
 export class InteractionSyllabifyComponent extends InteractionComponentDirective implements OnInit {
   parameters = input.required<SyllabifyParams>();
   options = signal<NumberedOption[]>([]);
+  selectedValues = signal<string>('');
 
   ngOnInit() {
-    const squareOptions: NumberedOption[] = Array.from(
-      { length: this.parameters().numberOfOptions },
-      (_, index) => ({
-        id: index,
-        text: (index + 1).toString()
-      })
-    );
+    const squareOptions = createNumberedOptions(this.parameters().numberOfOptions);
 
     this.options.set(squareOptions);
 
     /* Initialize the selectedValues with all 0 */
-    this.initializeBinarySelection(this.parameters().numberOfOptions);
+    this.selectedValues.set('0'.repeat(this.parameters().numberOfOptions));
   }
 
   onButtonClick(optionId: number): void {
-    this.toggleBinarySelection(optionId, this.parameters().variableId, 'SYLLABIFY_BUTTONS');
+    const newState = this.selectedValues().split('');
+    newState[optionId] = newState[optionId] === '1' ? '0' : '1';
+    this.selectedValues.set(newState.join(''));
+
+    const id = this.parameters().variableId || 'SYLLABIFY_BUTTONS';
+
+    const response: Response = {
+      id: id,
+      status: 'VALUE_CHANGED',
+      value: this.selectedValues()
+    };
+
+    this.responsesService.newResponses([response]);
+    this.responses.emit([response]);
   }
 
   isSelected(optionId: number): boolean {
-    return this.isSelectedAtIndex(optionId);
+    return this.selectedValues()[optionId] === '1';
   }
 }
